+++ src/main/java/org/dynjs/runtime/builtins/types/array/prototype/Filter.java
@@ -21,7 +21,7 @@ public class Filter extends AbstractNativeFunction {
     public Object call(ExecutionContext context, Object self, Object... args) {
         // 15.4.4.20
         JSObject o = Types.toObject(context, self);
-        int len = Types.toUint32(context, o.get(context, "length"));
+        long len = Types.toUint32(context, o.get(context, "length"));
 
         if (!(args[0] instanceof JSFunction)) {
             throw new ThrowException(context.createTypeError("callbackFn must be a function"));
@@ -38,7 +38,7 @@ public class Filter extends AbstractNativeFunction {
 
         int to = 0;
 
-        for (int k = 0; k < len; ++k) {
+        for (long k = 0; k < len; ++k) {
             boolean kPresent = o.hasProperty(context, "" + k);
             if (kPresent) {
                 final Object kValue = o.get(context, "" + k);
