+++ src/main/java/org/dynjs/runtime/builtins/types/array/prototype/Push.java
@@ -16,7 +16,7 @@ public class Push extends AbstractNativeFunction {
     public Object call(ExecutionContext context, Object self, Object... args) {
         // 15.4.4.2
         JSObject array = Types.toObject(context, self);
-        int n = Types.toUint32(context, array.get(context, "length"));
+        long n = Types.toUint32(context, array.get(context, "length"));
 
         for (Object each : args) {
             array.put(context, "" + n, each, true);
